immad/apps/substitution/aiida-sub-models/pyproject.toml
[build-system]
# build the package with [flit](https://flit.readthedocs.io)
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

[project]
# See https://www.python.org/dev/peps/pep-0621/
name = "aiida-sub-models"
dynamic = ["version"]   # read from aiida_sub_models/__init__.py
description = "Plugin for calling Substitution ML Models"
dependencies = [
    "aiida-core>=2.0,<3",
]

[project.entry-points."aiida.calculations"]
"sub-models" = "aiida_sub_models.calculations:TensorflowCalculation"

[project.entry-points."aiida.parsers"]
"sub-models" = "aiida_sub_models.parsers:TensorflowParser"

[tool.flit.module]
name = "aiida_sub_models"
